aws-strands/content_moderation_agent/pyproject.toml
[project]
name = "content-moderation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.4.9",
    "strands-agents>=1.1.0",
]
